Added hook for callback in case of failure to load image.
We've found that we need to be able to hook up some function in case the texture fails to load. In our case it's an API which can return a application/json in case there is no image which kicks off functions hooked up to the onerror property of the image.

See: http://stackoverflow.com/a/9815854/501017